pyproject.toml
[tool.poetry]
name = "led-cube-view"
version = "1.1.2"
description = "A QtPy widget to display a LED cube on a 3D graph and manipulate its LEDs."
authors = ["crash8229 <mu304007@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
packages = [{include = "led_cube_view"}]
homepage = "https://github.com/crash8229/led-cube-view"
repository = "https://github.com/crash8229/led-cube-view"
classifiers = ["Development Status :: 4 - Beta"]

[tool.poetry.dependencies]
python = "^3.10"
PyOpenGL = "^3.1.6"
numpy-stl = "^3.0.0"
pyqtgraph = "^0.13.1"
QtPy = "^2.3.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
mypy = "^0.982"
PySide6 = {version = "^6.3.1", python = ">=3.10,<3.12"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
